Set BPF_F_ADJ_ROOM_FIXED_GSO during vxlan decap correctly
The flag has been added to the mode flag where it is a noop, the GSO
size needs to be fixed as otherwise if packets are at the close to the
mtu the gso size will be increased and then dropped by tcp_gso_segment
check skb->len <= gso_size.
The packet is then retransmitted which causes delays.

This can be reproduced with following bpftraces and traffic with
appropriately sized packets but the exact conditions are not well
understood yet.

```
kprobe:tcp_gso_segment {
    $skb = (struct sk_buff*)arg0;
    $shinfo =(struct skb_shared_info*)($skb->head + $skb->end);
    if ($skb->len <= $shinfo->gso_size) {
        printf("DROP skb->len %ld gso_size %ld %d %d\n", $skb->len, $shinfo->gso_size,
              $shinfo->gso_type, ((struct napi_gro_cb*)$skb->cb)->count);
    }
}
kprobe:bpf_skb_adjust_room {
        $skb = (struct sk_buff*)arg0;
        if (((struct napi_gro_cb*)$skb->cb)->count > 1) {
                $shinfo =(struct skb_shared_info*)($skb->head + $skb->end);
                printf("adj skb->len %ld gso_size %ld len_diff %ld mode %u flags %lu\n", $skb->len, $shinfo->gso_size, (int32) arg1, (uint32) arg2, (uint32) arg3);
        }
}

> adj  skb->len 1466 gso_size 1348 len_diff -50 mode 1 flags 0
> DROP skb->len 1382 gso_size 1398 1027 2

```

Closes gh-11160