refactor: determine images based on the images release dirs
The release tool is now more integrated with individual components to dynamically construct images being built.
It add target to enable retrieving images built per dir and as a result we can reduce the numer of points where a component and/or its images need to be updated manually for release.