update GIT_VERSION for hashreleases
Hashrelease expect version in the format `vX.Y.Z-0.dev-N-gHHHHHHHHHHHH` or `vX.Y.Z-N-gHHHHHHHHHHHH`
where `vX.Y.Z` is either the expected release version (when the dev tag `0.dev` is a direct suffix`) or the latest release version,
`N` is the number of commits since the last tag and `HHHHHHHHHHHH` is the git hash.

While this can be achieved from running the below command:

```sh
git describe --tags --always --abbrev=12
```

However, on a commit that also has a tag, this will not include the `N-gHHHHHHHHHHHH`. Adding `--long` to ensure that images for hashreleases get tagged correctly.