Allow shimming of default location time for tests
This allows for easier comprisons of times for tests as two equal times with different locals aren't considered equal.

Note that there is an ENV variable TZ that can be used to update the local for the underlying time system. This proved unhelpful when comparing unset Time values (that don't have a local) with third party libraries that "parsed" empty time value strings (that would have a local).

So just allowing overriding the local for our package to match what we want from third party parsing proves much better.

Note that this is only for testing, and it's required to have a testing.T to set the default local.