CI: Cache the working copy and go build cache in GCS
Previously, we tried to save/restore the go build cache to
Semaphroe's cache.  however, this never worked due to go's use
of mtime comparison.

On branch builds, cache the working copy using tar so that
we save mtimes.  Then, for the "build" jobs, save off the go
build cache for each group of jobs as a second tarfile.
Different components build with different go options (CGO,
race detector, etc) so using a shared single cache is not optimal.

On all builds, restore the working copy cache and then do a
git fetch/checkout to get the intended commit. This is fast and
it maximises the effectiveness of the build cache.

On PR builds only, restore the build cache from the target
branch, this makes PR builds faster, saving 1.5-6 minutes per job.
Skip restoring the cache for jobs that don't build anything
(such as felix FV jobs).