CI: Tear down test VMs as soon as they're finished.
Previously, we ran all tests to completion before tearing
down any VMs. This meant that a lot of VMs were sat idle
twoards the end of the tests.  Refactor the test running logic
to remove idle time and tear down VMs as soon as they're done.

As a side effect, this staggers diags collection resulting
in less bandwidth contention.